refactor(update): expose helpers through test facades
